Moving work to agenda
