+++ server/src/tasks/apy.js
@@ -1,0 +1,7 @@
+const { calculate } = require('@utils/apy')
+
+const calculateApy = ({ walletAddress, tokenAddress }) => {
+  return calculate(walletAddress, tokenAddress)
+}
+
+module.exports = { calculateApy }
