+++ server/src/tasks/index.js
@@ -2,10 +2,12 @@ const events = require('./events')
 const actions = require('./actions')
 const alert = require('./alert')
 const tvl = require('./tvl')
+const apy = require('./apy')
 
 module.exports = {
   ...events,
   ...actions,
   ...alert,
-  ...tvl
+  ...tvl,
+  ...apy
 }
